webapi/webapi-sensors-w3c-tests/sensors/Magnetometer_browsing_context.https.html

<!DOCTYPE html>
<!--
Copyright (c) 2016 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-->

<meta charset="utf-8">
<title>Magnetometer Test: Sensor readings must only be available in the top-level browsing context</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Xiuqi Jiang" href="mailto:xiuqix.jiang@intel.com">
<link rel="help" href="https://w3c.github.io/magnetometer/">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="./resources/support-iframe.html" id="frame" style="display:none" sandbox="allow-scripts">
</iframe>
<script>

async_test(t => {
  window.onmessage = t.step_func(e => {
    assert_equals(e.data, "SecurityError");
    t.done();
  });
}, "throw a 'SecurityError' when firing sensor readings within iframes");

async_test(t => {
  let sensor = new Magnetometer();
  sensor.start();
  sensor.onactivate = t.step_func_done(() => {
    assert_not_equals(sensor.reading, null);
    let cachedReading = sensor.reading;
    let win = window.open('', '_blank');
    assert_equals(sensor.reading, cachedReading);
    win.close();
    sensor.stop();
  });
}, "sensor readings can not be fired on the background tab");

</script>
